common: move Range to its own file and add Iter
