+++ common/types.go
@@ -23,7 +23,6 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
 	"math/big"
 	"math/rand"
 	"reflect"
@@ -31,7 +30,6 @@ import (
 	"strings"
 
 	"github.com/ethereum/go-ethereum/common/hexutil"
-	"github.com/ethereum/go-ethereum/rlp"
 	"golang.org/x/crypto/sha3"
 )
 
@@ -487,84 +485,4 @@ func (b PrettyBytes) TerminalString() string {
 		return fmt.Sprintf("%x", b)
 	}
 	return fmt.Sprintf("%#x...%x (%dB)", b[:3], b[len(b)-3:], len(b))
-}
-
-type Range[T uint32 | uint64] struct {
-	first, afterLast T
-}
-
-func NewRange[T uint32 | uint64](first, count T) Range[T] {
-	return Range[T]{first, first + count}
-}
-
-func (r *Range[T]) EncodeRLP(w io.Writer) error {
-	if err := rlp.Encode(w, &r.first); err != nil {
-		return err
-	}
-	return rlp.Encode(w, &r.afterLast)
-}
-
-func (r *Range[T]) DecodeRLP(s *rlp.Stream) error {
-	if err := s.Decode(&r.first); err != nil {
-		return err
-	}
-	return s.Decode(&r.afterLast)
-}
-
-func (r Range[T]) First() T {
-	return r.first
-}
-
-func (r Range[T]) Last() T {
-	if r.first == r.afterLast {
-		panic("last item of zero length range is not allowed")
-	}
-	return r.afterLast - 1
-}
-
-func (r Range[T]) AfterLast() T {
-	return r.afterLast
-}
-
-func (r Range[T]) Count() T {
-	return r.afterLast - r.first
-}
-
-func (r Range[T]) IsEmpty() bool {
-	return r.first == r.afterLast
-}
-
-func (r Range[T]) Includes(v T) bool {
-	return v >= r.first && v < r.afterLast
-}
-
-func (r *Range[T]) SetFirst(v T) {
-	r.first = v
-	if r.afterLast < r.first {
-		r.afterLast = r.first
-	}
-}
-
-func (r *Range[T]) SetAfterLast(v T) {
-	r.afterLast = v
-	if r.afterLast < r.first {
-		r.first = r.afterLast
-	}
-}
-
-func (r *Range[T]) SetLast(v T) {
-	r.SetAfterLast(v + 1)
-}
-
-func (r Range[T]) Intersection(q Range[T]) Range[T] {
-	if r.first > q.first {
-		q.first = r.first
-	}
-	if r.afterLast < q.afterLast {
-		q.afterLast = r.afterLast
-	}
-	if q.first > q.afterLast {
-		return Range[T]{}
-	}
-	return q
 }
